refactor: simplify VersionedContentStore
